+++ src/renderer/components/LogEntries.vue
@@ -53,8 +53,10 @@
             onEdit: (item) => {
               edit(item);
             },
-            onArchiveDelete: (item) => {
-              refresh();
+            onArchiveDelete: (event, item) => {
+              if (event.startsWith('post-')) {
+                refresh();
+              }
             },
           }"
           v-bind:item="logEntry"